알라딘 api 연결
작성 페이지
- 책 정보 검색(제목, 저지, 이미지) 성공
- db 저장 에러 이슈 -> db에 모순 데이터 생성해 테스트중
북로그 메인 페이지
북로그 메인화면(홈) 테스트 완료 (북로그 이름, 내용, 책 정보, 날짜) 
-> 이미지 업로드 에러

+++ src/main/java/com/bookreum/domain/post/controller/CommentController.java
@@ -20,26 +20,26 @@ public class CommentController {
     // 댓글 생성
     @PostMapping
     public CommentDto.Response createComment(@RequestBody CommentDto.Request request) {
-        User user = User.builder().id(1L).nickname("테스터").build();
+        User user = User.builder().id(1).nickname("테스터").build();
         Post post = Post.builder().id(request.getPostId()).build();
         return commentService.createComment(request, user, post);
     }
 
     // 특정 게시글의 댓글 전체 조회 (본인 여부 확인)
     @GetMapping("/post/{postId}")
-    public List<CommentDto.Response> getCommentsByPost(@PathVariable Long postId, @RequestParam boolean isOwner) {
+    public List<CommentDto.Response> getCommentsByPost(@PathVariable Integer postId, @RequestParam boolean isOwner) {
         return commentService.getCommentsByPost(postId, isOwner);
     }
 
     // 댓글 삭제
     @DeleteMapping("/{id}")
-    public void deleteComment(@PathVariable Long id) {
+    public void deleteComment(@PathVariable Integer id) {
         commentService.deleteComment(id);
     }
 
     // 댓글 수정
     @PutMapping("/{id}")
-    public void updateComment(@PathVariable Long id, @RequestBody CommentDto.Request request) {
+    public void updateComment(@PathVariable Integer id, @RequestBody CommentDto.Request request) {
         commentService.updateComment(id, request.getContent());
     }
 }
