+++ src/main/java/com/bookreum/domain/post/controller/CommentHeartController.java
@@ -14,23 +14,23 @@ public class CommentHeartController {
 
     // 1. 댓글에 공감 토글 (누르면 추가, 다시 누르면 삭제)
     @PostMapping("/{commentId}/hearts")
-    public void toggleHeart(@PathVariable Long commentId) {
+    public void toggleHeart(@PathVariable Integer commentId) {
         // TODO: 인증 붙이면 user 추출해서 교체
-        User user = User.builder().id(1L).nickname("테스터").build();
+        User user = User.builder().id(1).nickname("테스터").build();
         commentHeartService.toggleHeart(commentId, user);
     }
 
     // 2. 댓글 공감 여부 확인
     @GetMapping("/{commentId}/hearts/exists")
-    public boolean hasHeart(@PathVariable Long commentId) {
+    public boolean hasHeart(@PathVariable Integer commentId) {
         // TODO: 인증 붙이면 user 추출해서 교체
-        User user = User.builder().id(1L).nickname("테스터").build();
+        User user = User.builder().id(1).nickname("테스터").build();
         return commentHeartService.hasHeart(commentId, user);
     }
 
     // 3. 댓글 공감 수 조회
     @GetMapping("/{commentId}/hearts/count")
-    public Long getHeartCount(@PathVariable Long commentId) {
+    public Long getHeartCount(@PathVariable Integer commentId) {
         return commentHeartService.getHeartCount(commentId);
     }
 }
