+++ src/main/java/com/bookreum/domain/post/repository/CommentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByPostId(Long postId);
-    long countByPost(Post post); 
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+    List<Comment> findByPostId(Integer postId);
+    long countByPost(Post post);
 }
